Make hardware code cross-platform.
