+++ Library/Homebrew/extend/os/hardware.rb
@@ -1,0 +1,5 @@
+if OS.mac?
+  require "extend/os/mac/hardware/cpu"
+elsif OS.linux?
+  require "extend/os/linux/hardware/cpu"
+end
