+++ Library/Homebrew/hardware.rb
@@ -1,60 +1,53 @@
 require "os"
 
-class Hardware
-  module CPU
-    extend self
+module Hardware
+  class CPU
     INTEL_32BIT_ARCHS = [:i386].freeze
     INTEL_64BIT_ARCHS = [:x86_64].freeze
     PPC_32BIT_ARCHS   = [:ppc, :ppc7400, :ppc7450, :ppc970].freeze
     PPC_64BIT_ARCHS   = [:ppc64].freeze
 
-    def type
-      :dunno
-    end
+    class << self
+      def type
+        :dunno
+      end
 
-    def family
-      :dunno
-    end
+      def family
+        :dunno
+      end
 
-    def cores
-      1
-    end
+      def cores
+        1
+      end
 
-    def bits
-      64
-    end
+      def bits
+        64
+      end
 
-    def is_32_bit?
-      bits == 32
-    end
+      def is_32_bit?
+        bits == 32
+      end
 
-    def is_64_bit?
-      bits == 64
-    end
+      def is_64_bit?
+        bits == 64
+      end
 
-    def intel?
-      type == :intel
-    end
+      def intel?
+        type == :intel
+      end
 
-    def ppc?
-      type == :ppc
-    end
+      def ppc?
+        type == :ppc
+      end
 
-    def features
-      []
-    end
+      def features
+        []
+      end
 
-    def feature?(name)
-      features.include?(name)
+      def feature?(name)
+        features.include?(name)
+      end
     end
-  end
-
-  if OS.mac?
-    require "os/mac/hardware"
-    CPU.extend MacCPUs
-  elsif OS.linux?
-    require "os/linux/hardware"
-    CPU.extend LinuxCPUs
   end
 
   def self.cores_as_words
@@ -79,3 +72,5 @@ class Hardware
     end
   end
 end
+
+require "extend/os/hardware"
