+++ Library/Homebrew/os/mac/architecture_list.rb
@@ -1,3 +1,5 @@
+require "extend/os/hardware"
+
 module ArchitectureListExtension
   # @private
   def fat?
